Added the ability to pass arguments to commands
We use strtok() to separate arguments with space and newline as the 
delimiters. Newline is a delimiter because fgets() includes the newline 
at the end, so we want to stop there for the last argument.

Added the macro ARGV_CHUNK_SZ which the size of the argv allocation 
chunks. Haven't added the ability to expand the argv array yet though. 
That is coming soon.