Move cwd to global scope and don't recalculate it after every loop
cwd often does not change when executing commands. In fact, it changes
only when calling cd. So, recalculating it every time we execute a
command is wasteful. Instead, only use it when either we are starting up
or when cd is called.

Also, while here, move cwd to global scope. It better resides there
because there is only one current working directory.

While here, check getcwd() for errors. If there is any, abort execution.
If this is not done, getcwd() returns NULL in case of an error. This
will lead to a segmentation fault.

This also reduces memory consumption a *little*. When cwd is a
subdirectory of HOME, the HOME's path is replaced by HOME_SYMBOL. But
the allocation is for the entire path, wasting a tiny bit of memory
(probably only a few bytes in most cases). Now, since we strdup(newcwd)
and then free(newcwd), there is no memory wastage.