We can handle signals now while waiting for the child to finish
This is achieved by a combination of getch() and SIGCHLD. After we fork
and exec, we call getch() what waits on terminal input. If the input
should raise a signal, then that signal is raised. Otherwise, the input
is ignored.

While getch() is sleeping and the child terminates, the parent receives
SIGCHLD which interrupts getch(). Then we can get out and return from
psh_exec().

I had to use some hacks to get it to work, and the code might not be
very portable, but I have commented about that, so check them out.